Solve task

+++ TaskSolution/HourlyEmployee.cpp
@@ -9,5 +9,5 @@ HourlyEmployee::~HourlyEmployee()
 
 double HourlyEmployee::average_monthly_wage()
 {
-	return (20.8 * (double)8 * this->emplFields.wage);
+	return (20.8 * 8 * this->emplFields.wage);
 }
